entry function worked and displayed in html
