+++ scripts/JournalDataProvider.js
