+++ scripts/main.js
@@ -1,3 +1,5 @@
 console.log("Welcome to the main module")
 
+import { EntryListComponent } from "./JournalEntryList.js"
 
+EntryListComponent()
